Preserve the base URI when fixupXmlBase is false
Fix #28

If `fixupXmlBase` is false, there’s nothing in the document to preserve the base
URI of an element when it’s nested inside other elements. In this case, we add a
magic attribute to keep track and then do a second top-to-bottom pass to remove
the magic attributes and set the base URIs correctly. There’s a time and space
cost, but there doesn’t seem to be any other solution.